first version of component-based view
